server/requirements.txt
# Core Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pydantic==2.9.2

# Anthropic (Claude 3.5 Sonnet)
anthropic==0.40.0

# Liquid Trading SDK
liquidtrading-python==0.1.3

# HTTP Client (for news ingestion)
httpx==0.27.2
aiohttp==3.11.10

# Async support
anyio==4.7.0

# Data validation & serialization
python-multipart==0.0.18

# Rate limiting
slowapi==0.1.9

# CORS
starlette==0.41.3

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.27.2
